Move person profile component and context
