+++ frontend/src/employee-frontend/App.tsx
@@ -27,7 +27,6 @@ import Header from './components/Header'
 import IncomeStatementPage from './components/IncomeStatementPage'
 import LoginPage from './components/LoginPage'
 import MobilePairingModal from './components/MobilePairingModal'
-import PersonProfile from './components/PersonProfile'
 import PersonalMobileDevicesPage from './components/PersonalMobileDevicesPage'
 import Reports from './components/Reports'
 import UnitFeaturesPage from './components/UnitFeaturesPage'
@@ -66,6 +65,7 @@ import InvoicesPage from './components/invoices/InvoicesPage'
 import MessagesPage from './components/messages/MessagesPage'
 import OutOfOfficePage from './components/out-of-office/OutOfOfficePage'
 import PaymentsPage from './components/payments/PaymentsPage'
+import PersonProfile from './components/person-profile/PersonProfile'
 import Search from './components/person-search/Search'
 import PlacementDraftPage from './components/placement-draft/PlacementDraft'
 import PlacementToolPage from './components/placement-tool/PlacementToolPage'
