+++ frontend/src/employee-frontend/components/child-information/Guardians.tsx
@@ -30,8 +30,8 @@ import {
 import { ChildContext } from '../../state'
 import { useTranslation } from '../../state/i18n'
 import { formatName } from '../../utils'
-import { NameTd } from '../PersonProfile'
 import { renderResult } from '../async-rendering'
+import { NameTd } from '../person-profile/PersonProfile'
 
 const getPersonGuardiansResult = wrapResult(getPersonGuardians)
 const updateGuardianEvakaRightsResult = wrapResult(updateGuardianEvakaRights)
