+++ frontend/src/employee-frontend/components/person-profile/PersonFridgePartner.tsx
@@ -27,11 +27,12 @@ import {
   retryPartnership
 } from '../../generated/api-clients/pis'
 import { useTranslation } from '../../state/i18n'
-import { PersonContext } from '../../state/person'
 import { UIContext } from '../../state/ui'
 import { formatName } from '../../utils'
-import { ButtonsTd, DateTd, NameTd } from '../PersonProfile'
 import { renderResult } from '../async-rendering'
+
+import { ButtonsTd, DateTd, NameTd } from './PersonProfile'
+import { PersonContext } from './state'
 
 const getPartnershipsResult = wrapResult(getPartnerships)
 const deletePartnershipResult = wrapResult(deletePartnership)
